+++ Converter.java
@@ -16,6 +16,16 @@ public class Converter {
         return result;
     }
 
+    public static int bin2dec(boolean[] bin) {
+        int sum = 0;
+        for (int i=0; i<bin.length; i++) {
+            sum += (bin[i])
+                    ? intExponentiation(2,i) // 현재 자리수가 true이면 2의 i승을 더함.
+                    : 0; // 아니면 0을 더함.
+        }
+        return sum;
+    }
+
     private static int intExponentiation(int base, int exponent) {
         int result = 1;
         for (int i=0; i<exponent; i++) {